#3: Moved tests to their own files for easier viewing.
