+++ src/__tests__/to-have-attributes.js
@@ -1,0 +1,38 @@
+import '../extend-expect'
+import {plugins} from 'pretty-format'
+import {render} from './helpers/test-utils'
+
+expect.addSnapshotSerializer(plugins.ConvertAnsi)
+
+test('.toHaveAttribute', () => {
+  const {queryByTestId} = render(`
+    <button data-testid="ok-button" type="submit" disabled>
+      OK
+    </button>
+  `)
+
+  expect(queryByTestId('ok-button')).toHaveAttribute('disabled')
+  expect(queryByTestId('ok-button')).toHaveAttribute('type')
+  expect(queryByTestId('ok-button')).not.toHaveAttribute('class')
+  expect(queryByTestId('ok-button')).toHaveAttribute('type', 'submit')
+  expect(queryByTestId('ok-button')).not.toHaveAttribute('type', 'button')
+
+  expect(() =>
+    expect(queryByTestId('ok-button')).not.toHaveAttribute('disabled'),
+  ).toThrowError()
+  expect(() =>
+    expect(queryByTestId('ok-button')).not.toHaveAttribute('type'),
+  ).toThrowError()
+  expect(() =>
+    expect(queryByTestId('ok-button')).toHaveAttribute('class'),
+  ).toThrowError()
+  expect(() =>
+    expect(queryByTestId('ok-button')).not.toHaveAttribute('type', 'submit'),
+  ).toThrowError()
+  expect(() =>
+    expect(queryByTestId('ok-button')).toHaveAttribute('type', 'button'),
+  ).toThrowError()
+  expect(() =>
+    expect({thisIsNot: 'an html element'}).not.toHaveAttribute(),
+  ).toThrowError()
+})
