+++ src/__tests__/to-have-class.js
@@ -1,0 +1,48 @@
+import '../extend-expect'
+import {plugins} from 'pretty-format'
+import {render} from './helpers/test-utils'
+
+expect.addSnapshotSerializer(plugins.ConvertAnsi)
+
+test('.toHaveClass', () => {
+  const {queryByTestId} = render(`
+    <div>
+      <button data-testid="delete-button" class="btn extra btn-danger">
+        Delete item
+      </button>
+      <button data-testid="cancel-button">
+        Cancel
+      </button>
+    </div>
+  `)
+
+  expect(queryByTestId('delete-button')).toHaveClass('btn')
+  expect(queryByTestId('delete-button')).toHaveClass('btn-danger')
+  expect(queryByTestId('delete-button')).toHaveClass('extra')
+  expect(queryByTestId('delete-button')).not.toHaveClass('xtra')
+  expect(queryByTestId('delete-button')).toHaveClass('btn btn-danger')
+  expect(queryByTestId('delete-button')).not.toHaveClass('btn-link')
+  expect(queryByTestId('cancel-button')).not.toHaveClass('btn-danger')
+
+  expect(() =>
+    expect(queryByTestId('delete-button')).not.toHaveClass('btn'),
+  ).toThrowError()
+  expect(() =>
+    expect(queryByTestId('delete-button')).not.toHaveClass('btn-danger'),
+  ).toThrowError()
+  expect(() =>
+    expect(queryByTestId('delete-button')).not.toHaveClass('extra'),
+  ).toThrowError()
+  expect(() =>
+    expect(queryByTestId('delete-button')).toHaveClass('xtra'),
+  ).toThrowError()
+  expect(() =>
+    expect(queryByTestId('delete-button')).not.toHaveClass('btn btn-danger'),
+  ).toThrowError()
+  expect(() =>
+    expect(queryByTestId('delete-button')).toHaveClass('btn-link'),
+  ).toThrowError()
+  expect(() =>
+    expect(queryByTestId('cancel-button')).toHaveClass('btn-danger'),
+  ).toThrowError()
+})
